+++ db/schema.rb
@@ -89,7 +89,7 @@ ActiveRecord::Schema[7.1].define(version: 2024_02_07_093611) do
   create_table "stats", force: :cascade do |t|
     t.bigint "item_id"
     t.string "stat", null: false
-    t.integer "amount", null: false
+    t.decimal "amount", null: false
     t.index ["item_id"], name: "index_stats_on_item_id"
     t.index ["stat"], name: "index_stats_on_stat"
   end
